requirements.txt
# Core dependencies
langchain>=0.1.0
langchain-community>=0.0.13
langchain-openai>=0.0.2
openai>=1.6.0
tiktoken>=0.5.2
faiss-cpu>=1.7.4
sentence-transformers>=2.2.2
python-dotenv>=1.0.0
pydantic>=2.5.0

# Text processing
spacy>=3.7.2
nltk>=3.8.1

# Optional - UI
streamlit>=1.29.0

# Optional - Web Search
beautifulsoup4>=4.12.2
requests>=2.31.0
# Uncomment if implementing search API option
# google-search-results>=2.4.2  # SerpAPI